Moved user assigned workevent related classes from auth to user
