+++ wman-backend/Wman.Logic/Classes/AuthLogic.cs
@@ -242,28 +242,5 @@ namespace Wman.Logic.Classes
             var users = await this.userManager.GetUsersInRoleAsync(roleId);
             return users.ToList();
         }
-
-        public async Task<IEnumerable<AssignedEventDTO>> JobsOfUser(string username)
-        {
-            var selectedUser = await userManager.Users
-                .Where(x => x.UserName == username)
-                .Include(y => y.WorkEvents)
-                .ThenInclude(z => z.Address)
-                .AsNoTracking()
-                .SingleOrDefaultAsync();
-            if (selectedUser == null)
-            {
-                throw new ArgumentException("User not found!");
-            }
-            var output = selectedUser.WorkEvents;
-            if (output.Count() == 0)
-            {
-                throw new InvalidOperationException("User has no assigned jobs! ");
-            }
-            ;
-            var testResult = mapper.Map<IEnumerable<AssignedEventDTO>>(output);
-
-            return testResult;
-        }
     }
 }
