+++ wman-backend/Wman.Logic/Interfaces/IAuthLogic.cs
@@ -42,7 +42,5 @@ namespace Wman.Logic.Interfaces
 
         Task<bool> HasRoleByName(string userName, string role);
 
-        Task<IEnumerable<AssignedEventDTO>> JobsOfUser(string username);
-
     }
 }
