+++ wman-backend/Wman.Logic/Interfaces/IUserLogic.cs
@@ -11,5 +11,7 @@ namespace Wman.Logic.Interfaces
     {
         public Task<IEnumerable<WorkloadDTO>> getWorkLoads(IEnumerable<string> usernames);
         public Task<IEnumerable<WorkloadDTO>> getWorkLoads();
+
+        Task<IEnumerable<AssignedEventDTO>> JobsOfUser(string username);
     }
 }
